factor out reallocation form into component
