+++ src/pages/court-cases/[courtCaseId]/reallocate.tsx
@@ -1,23 +1,19 @@
-import { forces } from "@moj-bichard7-developers/bichard7-next-data"
-import ButtonsGroup from "components/ButtonsGroup"
 import ConditionalRender from "components/ConditionalRender"
 import HeaderContainer from "components/Header/HeaderContainer"
 import HeaderRow from "components/Header/HeaderRow"
 import Layout from "components/Layout"
-import { MAX_NOTE_LENGTH } from "config"
 import { CurrentUserContext, CurrentUserContextType } from "context/CurrentUserContext"
-import { BackLink, Button, Fieldset, FormGroup, Heading, HintText, Label, Link, Select, TextArea } from "govuk-react"
+import { BackLink, Heading } from "govuk-react"
 import { withAuthentication, withMultipleServerSideProps } from "middleware"
 import type { GetServerSidePropsContext, GetServerSidePropsResult, NextPage } from "next"
 import Head from "next/head"
 import { useRouter } from "next/router"
 import { ParsedUrlQuery } from "querystring"
-import { FormEventHandler, useState } from "react"
+import { useState } from "react"
 import { courtCaseToDisplayFullCourtCaseDto } from "services/dto/courtCaseDto"
 import { userToDisplayFullUserDto } from "services/dto/userDto"
 import getCourtCaseByOrganisationUnit from "services/getCourtCaseByOrganisationUnit"
 import getDataSource from "services/getDataSource"
-import getForcesForReallocation from "services/getForcesForReallocation"
 import reallocateCourtCase from "services/reallocateCourtCase"
 import AuthenticationServerSidePropsContext from "types/AuthenticationServerSidePropsContext"
 import { isError } from "types/Result"
@@ -26,13 +22,12 @@ import { DisplayFullUser } from "types/display/Users"
 import forbidden from "utils/forbidden"
 import { isPost } from "utils/http"
 import redirectTo from "utils/redirectTo"
-import { useCustomStyles } from "../../../../styles/customStyles"
-import Form from "../../../components/Form"
 import withCsrf from "../../../middleware/withCsrf/withCsrf"
 import CsrfServerSidePropsContext from "../../../types/CsrfServerSidePropsContext"
 import { CourtCaseContext, CourtCaseContextType } from "context/CourtCaseContext"
 import { CsrfTokenContext, CsrfTokenContextType } from "context/CsrfTokenContext"
-import { UserNotes } from "features/CourtCaseDetails/Tabs/Panels/Notes/UserNotes"
+import { UserNotesTable } from "features/CourtCaseDetails/Tabs/Panels/Notes/UserNotesTable"
+import ReallocationNotesForm from "components/ReallocationNotesForm"
 
 export const getServerSideProps = withMultipleServerSideProps(
   withAuthentication,
@@ -106,14 +101,7 @@ const ReallocateCasePage: NextPage<Props> = ({
   csrfToken,
   previousPath
 }: Props) => {
-  const [noteRemainingLength, setNoteRemainingLength] = useState(MAX_NOTE_LENGTH)
-  const classes = useCustomStyles()
   const { basePath } = useRouter()
-  const currentForce = forces.find((force) => force.code === courtCase.orgForPoliceFilter?.substring(0, 2))
-  const forcesForReallocation = getForcesForReallocation(currentForce?.code)
-  const handleOnNoteChange: FormEventHandler<HTMLTextAreaElement> = (event) => {
-    setNoteRemainingLength(MAX_NOTE_LENGTH - event.currentTarget.value.length)
-  }
   const [currentUserContext] = useState<CurrentUserContextType>({ currentUser: user })
   const [courtCaseContext] = useState<CourtCaseContextType>({ courtCase: courtCase })
   const [csrfTokenContext] = useState<CsrfTokenContextType>({ csrfToken })
@@ -149,44 +137,8 @@ const ReallocateCasePage: NextPage<Props> = ({
               </ConditionalRender>
               <ConditionalRender isRendered={!lockedByAnotherUser}>
                 <div className="govuk-grid-row">
-                  <Form className="govuk-grid-column-one-half" method="POST" action="#" csrfToken={csrfToken}>
-                    <Fieldset>
-                      <FormGroup>
-                        <Label>{"Current force owner"}</Label>
-                        <span>{`${currentForce?.code} - ${currentForce?.name}`}</span>
-                      </FormGroup>
-                      <FormGroup>
-                        <Label>{"New force owner"}</Label>
-                        <Select input={{ name: "force" }} label={""}>
-                          {forcesForReallocation.map(({ code, name }) => (
-                            <option key={code} value={code}>
-                              {`${code} - ${name}`}
-                            </option>
-                          ))}
-                        </Select>
-                      </FormGroup>
-                      <FormGroup>
-                        <Label>{"Add a note (optional)"}</Label>
-                        <HintText className={classes["no-margin-bottom"]}>
-                          {"Input reason for case reallocation"}
-                        </HintText>
-                        <TextArea
-                          input={{ name: "note", rows: 5, maxLength: MAX_NOTE_LENGTH, onInput: handleOnNoteChange }}
-                        >
-                          {""}
-                        </TextArea>
-                        <HintText>{`You have ${noteRemainingLength} characters remaining`}</HintText>
-                      </FormGroup>
-
-                      <ButtonsGroup>
-                        <Button id="Reallocate" type="submit">
-                          {"Reallocate"}
-                        </Button>
-                        <Link href={backLink}>{"Cancel"}</Link>
-                      </ButtonsGroup>
-                    </Fieldset>
-                  </Form>
-                  <UserNotes className="govuk-grid-column-one-half" />
+                  <ReallocationNotesForm csrfToken={csrfToken} courtCase={courtCase} backLink={backLink} />
+                  <UserNotesTable className="govuk-grid-column-one-half" />
                 </div>
               </ConditionalRender>
             </Layout>
